Split out manipulationstation demo
